+++ src/components/order/OrderList.tsx
@@ -1,19 +1,6 @@
-import {
-  Box,
-  Button,
-  Center,
-  Flex,
-  FormControl,
-  HStack,
-  Input,
-  Table,
-  Tbody,
-  Td,
-  Text,
-  Th,
-  Thead,
-  Tr,
-} from '@chakra-ui/react';
+import { Box, Button, Flex, Table, Tbody, Td, Text, Th, Thead, Tr } from '@chakra-ui/react';
+import OrderSearch from 'components/order/OrderSearch';
+import Paging from 'components/order/Paging';
 import { useSearchParams } from 'react-router-dom';
 import { Order } from 'types/Order';
 
@@ -21,57 +8,34 @@ interface OrderListParam {
   orderLists: Order[];
 }
 
+const filterDatasByStatus = (datas: Order[], filter: string | null) => {
+  return filter !== null ? datas.filter(order => '' + order.status === filter) : datas;
+};
+
+const searchDatasByCustomerName = (datas: Order[], search: string | null) => {
+  return search !== null ? datas.filter(order => order.customer_name === search) : datas;
+};
+
 const pageSize = 50;
-const pagingOffset = 10;
 
 const OrderList = ({ orderLists }: OrderListParam) => {
   const [searchParams, setSearchParams] = useSearchParams();
   const sortKey: string[] = searchParams.getAll('sort');
   const filter = searchParams.get('filter');
   const search = searchParams.get('search');
-  let viewDatas =
-    filter !== null ? orderLists.filter(order => '' + order.status === filter) : orderLists;
-  viewDatas =
-    search !== null ? viewDatas.filter(order => order.customer_name === search) : viewDatas;
+
+  let viewDatas = filterDatasByStatus(orderLists, filter);
+  viewDatas = searchDatasByCustomerName(viewDatas, search);
+
   sortKey.forEach(sortKey => {
     viewDatas.sort((a, b) =>
       (sortKey === 'id' ? a.id > b.id : a.transaction_time > b.transaction_time) ? -1 : 1,
     );
   });
+
   const totalOrders = viewDatas ? viewDatas.length - 1 : 0;
   const maxIndex = Math.floor(totalOrders / pageSize) - 1;
   const pagingIndex = parseInt(searchParams.get('page') ?? '1') - 1;
-  const startPageIndex = pagingIndex - ((pagingIndex - 1) % 10) - 1;
-  const lastPageIndex =
-    maxIndex >= startPageIndex + pagingOffset ? startPageIndex + pagingOffset : maxIndex + 1;
-
-  const handlePageMove = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-    const {
-      currentTarget: { value },
-    } = e;
-    searchParams.set('page', value);
-    setSearchParams(searchParams);
-  };
-
-  const handlePagePrevIndexs = () => {
-    setSearchParams(searchParams => {
-      searchParams.set('page', (pagingIndex <= pagingOffset ? 1 : pagingIndex - 10).toString());
-      return searchParams;
-    });
-  };
-
-  const handlePageNextIndexs = () => {
-    setSearchParams(searchParams => {
-      searchParams.set(
-        'page',
-        (maxIndex >= pagingIndex + pagingOffset
-          ? pagingIndex + pagingOffset
-          : maxIndex + 1
-        ).toString(),
-      );
-      return searchParams;
-    });
-  };
 
   const setSortKey = (key: string) => {
     const newSortKey = sortKey.includes(key)
@@ -95,18 +59,6 @@ const OrderList = ({ orderLists }: OrderListParam) => {
     });
   };
 
-  const setSearching = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    const searchValue = e.currentTarget.querySelector('input')?.value;
-    setSearchParams(searchParams => {
-      searchParams.delete('search');
-      if (searchValue) {
-        searchParams.set('search', searchValue);
-      }
-      return searchParams;
-    });
-  };
-
   return (
     <>
       <Flex
@@ -123,14 +75,18 @@ const OrderList = ({ orderLists }: OrderListParam) => {
         <Box display={'flex'} gap={2} alignItems='center'>
           <Text>정렬</Text>
           <Button
+            borderWidth='1px'
             onClick={() => setSortKey('id')}
-            backgroundColor={sortKey.includes('id') ? 'gray.300' : 'gray.100'}
+            backgroundColor={sortKey.includes('id') ? 'gray.400' : 'gray.100'}
+            name='id'
           >
             주문번호
           </Button>
           <Button
+            borderWidth='1px'
             onClick={() => setSortKey('transaction_time')}
-            backgroundColor={sortKey.includes('transaction_time') ? 'gray.300' : 'gray.100'}
+            name='transaction_time'
+            backgroundColor={sortKey.includes('transaction_time') ? 'gray.400' : 'gray.100'}
           >
             거래일 & 거래시간
           </Button>
@@ -138,25 +94,22 @@ const OrderList = ({ orderLists }: OrderListParam) => {
         <Box display={'flex'} gap={2} alignItems='center'>
           <Text>주문 처리 상태</Text>
           <Button
+            borderWidth='1px'
             onClick={() => setFiltering(true)}
-            backgroundColor={filter === 'true' ? 'gray.300' : 'gray.100'}
+            backgroundColor={filter === 'true' ? 'gray.400' : 'gray.100'}
           >
             처리
           </Button>
           <Button
+            borderWidth='1px'
             onClick={() => setFiltering(false)}
-            backgroundColor={filter === 'false' ? 'gray.300' : 'gray.100'}
+            backgroundColor={filter === 'false' ? 'gray.400' : 'gray.100'}
           >
             비처리
           </Button>
         </Box>
       </Flex>
-      <form onSubmit={setSearching}>
-        <FormControl pb={5} pr={5} display={'flex'} gap={2} justifyContent='flex-end'>
-          <Input placeholder='고객 이름을 입력해주세요' w='auto' name='searchCustomer' />
-          <Button type='submit'>검색</Button>
-        </FormControl>
-      </form>
+      <OrderSearch />
       <Box pb={5} pr={5}>
         <Text textAlign={'right'}>주문 건 : {viewDatas.length} 건</Text>
       </Box>
@@ -191,27 +144,7 @@ const OrderList = ({ orderLists }: OrderListParam) => {
         </Tbody>
       </Table>
 
-      {viewDatas.length > 0 && (
-        <Center>
-          <Box padding={7}>
-            <HStack>
-              <Button isDisabled={maxIndex < 10} onClick={handlePagePrevIndexs}>
-                &lt;
-              </Button>
-              {Array.from(new Array(maxIndex + 1), (_, key) => key + 1)
-                .slice(startPageIndex, lastPageIndex)
-                .map(value => (
-                  <Button key={value} value={value} onClick={handlePageMove}>
-                    {value}
-                  </Button>
-                ))}
-              <Button isDisabled={maxIndex < 10} onClick={handlePageNextIndexs}>
-                &gt;
-              </Button>
-            </HStack>
-          </Box>
-        </Center>
-      )}
+      {viewDatas.length > 0 && <Paging maxIndex={maxIndex} pagingIndex={pagingIndex} />}
     </>
   );
 };
